Fix type and add switch default for warnings
Changed window_size_ type in SimulatedAnnealing.h to unsigned long long for consistency. Added default case to switch statements in EnergyUtility.cpp to resolve -Wswitch-default compiler warnings.

+++ lmc/mc/include/SimulatedAnnealing.h
@@ -50,7 +50,7 @@ class SimulatedAnnealing : public McAbstract {
   // reheat thresholds (absolute steps)
   const unsigned long long int reheat_trigger_steps_;   // steps without improvement to trigger reheat
   const unsigned long long int reheat_cooldown_steps_;  // minimal interval between reheats
-  const unsigned int window_size_;                      // sliding window size for acceptance ratio
+  const unsigned long long int window_size_;            // sliding window size for acceptance ratio
 
   // (min, max) targets at early and late progress
   const double acc_max_{0.50};
